requirements.txt
tensorflow
pillow
scipy
cikit-learn
opencv-python